+++ src/proof/alethe/alethe_proof_rule.cpp
@@ -118,6 +118,23 @@ const char* aletheRuleToString(AletheRule id)
     case AletheRule::SYMM: return "symm";
     case AletheRule::NOT_SYMM: return "not_symm";
     case AletheRule::REORDERING: return "reordering";
+    case AletheRule::BV_BITBLAST_STEP_VAR: return "bv_bitblast_step_var";
+    case AletheRule::BV_BITBLAST_STEP_BVAND: return "bv_bitblast_step_bvand";
+    case AletheRule::BV_BITBLAST_STEP_BVOR: return "bv_bitblast_step_bvor";
+    case AletheRule::BV_BITBLAST_STEP_BVXOR: return "bv_bitblast_step_bvxor";
+    case AletheRule::BV_BITBLAST_STEP_BVXNOR: return "bv_bitblast_step_bvxnor";
+    case AletheRule::BV_BITBLAST_STEP_BVNOT: return "bv_bitblast_step_bvnot";
+    case AletheRule::BV_BITBLAST_STEP_BVADD: return "bv_bitblast_step_bvadd";
+    case AletheRule::BV_BITBLAST_STEP_BVNEG: return "bv_bitblast_step_bvneg";
+    case AletheRule::BV_BITBLAST_STEP_BVMULT: return "bv_bitblast_step_bvmult";
+    case AletheRule::BV_BITBLAST_STEP_BVULE: return "bv_bitblast_step_bvule";
+    case AletheRule::BV_BITBLAST_STEP_BVULT: return "bv_bitblast_step_bvult";
+    case AletheRule::BV_BITBLAST_STEP_EXTRACT:
+      return "bv_bitblast_step_extract";
+    case AletheRule::BV_BITBLAST_STEP_BVEQUAL:
+      return "bv_bitblast_step_bvequal";
+    case AletheRule::BV_BITBLAST_STEP_CONCAT: return "bv_bitblast_step_concat";
+    case AletheRule::BV_BITBLAST_STEP_CONST: return "bv_bitblast_step_const";
     //================================================= Undefined rule
     case AletheRule::UNDEFINED: return "undefined";
     default: return "?";