+++ src/theory/uf/equality_engine.cpp
@@ -1088,7 +1088,10 @@ void EqualityEngine::buildEqConclusion(EqualityNodeId id1,
   if (!eq[0].isNull() && !eq[1].isNull())
   {
     eqp->d_node = eq[0].eqNode(eq[1]);
+    Trace("equality") << "buildEqConclusion: Built equality " << eqp->d_node << "\n";
+    return;
   }
+  Trace("equality") << "buildEqConclusion: Did not build equality\n";
 }
 
 void EqualityEngine::explainEquality(TNode t1, TNode t2, bool polarity,