+++ test/unit/theory/theory_bv_black.cpp
@@ -58,6 +58,7 @@ class TestTheoryBlackBv : public TestApi
   void test_signed_overflow(cvc5::Kind kind, cvc5::Kind kindo)
   {
     d_solver.setOption("incremental", "true");
+    d_solver.setOption("produce-models", "true");
     for (uint32_t w = 1; w < 8; ++w)
     {
       d_solver.push();
@@ -67,18 +68,33 @@ class TestTheoryBlackBv : public TestApi
       Op sext = d_solver.mkOp(BITVECTOR_SIGN_EXTEND, {w});
       Term zx = d_solver.mkTerm(sext, {x});
       Term zy = d_solver.mkTerm(sext, {y});
-      Term mul = d_solver.mkTerm(kind, {zx, zy});
+      Term op = d_solver.mkTerm(kind, {zx, zy});
 
       Term max = d_solver.mkBitVector(
           2 * w, static_cast<uint32_t>(std::pow(2, w - 1)));
       Term min = d_solver.mkTerm(BITVECTOR_NEG, {max});
 
       Term lhs = d_solver.mkTerm(OR,
-                                 {d_solver.mkTerm(BITVECTOR_SLT, {mul, min}),
-                                  d_solver.mkTerm(BITVECTOR_SGE, {mul, max})});
+                                 {d_solver.mkTerm(BITVECTOR_SLT, {op, min}),
+                                  d_solver.mkTerm(BITVECTOR_SGE, {op, max})});
+      if (kind == BITVECTOR_SDIV)
+      {
+        lhs = d_solver.mkTerm(
+            AND,
+            {d_solver.mkTerm(DISTINCT, {y, d_solver.mkBitVector(w)}), lhs});
+      }
       Term rhs = d_solver.mkTerm(kindo, {x, y});
       Term eq = d_solver.mkTerm(DISTINCT, {lhs, rhs});
       d_solver.assertFormula(eq);
+      if (d_solver.checkSat().isSat())
+      {
+        std::cout << "x: " << d_solver.getValue(x) << std::endl;
+        std::cout << "y: " << d_solver.getValue(y) << std::endl;
+        std::cout << "op: " << d_solver.getValue(op) << std::endl;
+        std::cout << "max: " << d_solver.getValue(max) << std::endl;
+        std::cout << "lhs: " << d_solver.getValue(lhs) << std::endl;
+        std::cout << "rhs: " << d_solver.getValue(rhs) << std::endl;
+      }
       ASSERT_TRUE(d_solver.checkSat().isUnsat());
       d_solver.pop();
     }
@@ -113,6 +129,11 @@ TEST_F(TestTheoryBlackBv, usubo)
 TEST_F(TestTheoryBlackBv, ssubo)
 {
   test_signed_overflow(BITVECTOR_SUB, BITVECTOR_SSUBO);
+}
+
+TEST_F(TestTheoryBlackBv, sdivo)
+{
+  test_signed_overflow(BITVECTOR_SDIV, BITVECTOR_SDIVO);
 }
 
 TEST_F(TestTheoryBlackBv, reg8361)